Wip more restructuring.
