+++ style/index.css
@@ -5,25 +5,31 @@
 .dask-DaskSidebar {
   background: var(--jp-layout-color1);
   color: var(--jp-ui-font-color1);
-  display: flex;
-  flex-direction: column;
   font-size: var(--jp-ui-font-size1);
 }
 
 .dask-DashboardListing-list {
-  margin: 24px;
+  display: flex;
+  flex-direction: row;
+  flex-wrap: wrap;
+  justify-content: center;
+  margin: 8px;
   padding: 0;
   list-style-type: none;
 }
 
 .dask-DashboardListing-item {
-  display: flex;
-  margin: 4px 0 4px 0;
-  width: 100%;
+  margin: 4px;
+  white-space: nowrap;
+}
+
+.dask-DashboardListing-item button {
+  font-size: var(--jp-ui-font-size0);
+  line-height: 24px;
+  padding: 0px 8px;
 }
 
 .dask-DashboardListing-item button.jp-mod-styled.jp-mod-accept {
-  flex: 1 1 auto;
   height: var(--dask-launch-button-height);
   color: #000000;
   background-color: #ecb172;
